user_db_photo_amend_bigint

+++ database/migrations/2018_01_24_143449_create_table_user.php
@@ -19,7 +19,7 @@ class CreateTableUser extends Migration
             $table->string('password');
             $table->string('email')->unique();
             $table->text('data')->nullable();
-            $table->integer('phone')->unique();
+            $table->bigInteger('phone')->unique();
             $table->string('permission')->nullable();
             $table->timestamps();
         });
